refactor: renamed dismissibles
