+++ examples/lib/src/full_screen_dialog_manager/screens/full_screen_dialog_manager_basic_usage_screen.dart
@@ -46,7 +46,7 @@ final _backgroundAnimators = <String, FullScreenBackgroundAnimator>{
 };
 
 const _dismissibles = <String, FullScreenDismissible>{
-  _dismissibleGesture: FullScreenDismissible.gesture(),
+  _dismissibleGesture: FullScreenDismissible.tap(),
   _dismissibleNone: FullScreenDismissible.none(),
 };
