+++ packages/full_screen_dialog_manager/test/src/dismissible/full_screen_dismissible_test.dart
@@ -4,7 +4,7 @@ import 'package:full_screen_dialog_manager/full_screen_dialog_manager.dart';
 void main() {
   test('create all', () {
     expect(
-      () => FullScreenDismissible.gesture(
+      () => FullScreenDismissible.tap(
         onDismissed: () {},
       ),
       returnsNormally,
